backend/requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# HTTP requests and web scraping
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.4.0

# Database (optional - for full functionality)
asyncpg==0.30.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# NLP and text processing
textblob==0.17.1
langdetect==1.0.9
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0

# Data processing
pandas==2.1.3
numpy==1.25.2

# Additional utilities
aiofiles==23.2.1
python-multipart==0.0.6
lxml==4.9.3
